Copy standard src/
This is a larger commit than usual for the copy/paste parts of a new project, because I'm testing being able to do a simple copy/paste of the src/ directory and main scripts now that configuration has been moved to its own folder.
The actual configuration will come in a future commit.